==> Python/ML_basic/4.Logistic Regression/4.Logistic_Regression_with_sklearn.ipynb <==
import pandas as pd

# !wget http://www-stat.wharton.upenn.edu/~waterman/DataSets/uva.txt

# data_url= "http://www-stat.wharton.upenn.edu/~waterman/DataSets/uva.txt"
data_url= "data/uva.txt"
dataframe = pd.read_table(data_url)

dataframe[:5]

dataframe.shape

dataframe.dtypes

dataframe["Marital Status"].unique()

# 필요없는 행 제거
dataframe.pop('who')   # dataframe = dataframe.ix[:, 1:]
dataframe.pop('Country')
dataframe.pop('Years on Internet')

dataframe.dtypes

# "Gender"을 category 변수로 data type 변경
dataframe["Gender"] = dataframe["Gender"].astype("category")
dataframe.dtypes

for col in ['Household Income', 'Sexual Preference', 'Education Attainment', 'Major Occupation', "Marital Status"]:
    dataframe[col] = dataframe[col].astype('category')
dataframe.dtypes    

dataframe_modified = pd.get_dummies(dataframe)
dataframe_modified[:5]

dataframe_modified.shape

dataframe_modified.isnull().sum()

# "Age"가 결측치인 값에는 "Age" 값들의 평균으로 채워라
dataframe_modified.loc[
    pd.isnull(dataframe_modified['Age']), "Age"] = dataframe_modified['Age'].mean()

dataframe_modified['Age'][pd.isnull(dataframe_modified['Age'])]

dataframe_modified.head()

x_data = dataframe_modified.iloc[:, 1:].values
y_data = dataframe_modified.iloc[:, 0].values.reshape(-1, 1)
y_data.shape, x_data.shape

from sklearn import preprocessing   # Min-Max Standardzation

min_max_scaler = preprocessing.MinMaxScaler()
x_data = min_max_scaler.fit_transform(x_data)

import numpy as np 

# train, test set 나누기 위한 index 리스트 생성
training_idx = np.random.randint(y_data.shape[0], size=int(y_data.shape[0] * 0.8))
test_idx = np.random.randint(y_data.shape[0], size=int(y_data.shape[0] * 0.2))

x_training, x_test = x_data[training_idx,:], x_data[test_idx,:]
y_training, y_test = y_data[training_idx,:], y_data[test_idx,:]

x_training.shape, x_test.shape

from sklearn import linear_model, datasets

logreg = linear_model.LogisticRegression(fit_intercept=True, solver='liblinear')
logreg.fit(x_training, y_training.flatten())

logreg.predict(x_test[:5])

logreg.predict_proba(x_test[:5])

x_test.shape

y_test.shape

logreg.predict(x_test) == y_test.flatten()

# Accuracy
sum(logreg.predict(x_test) == y_test.flatten()) / len(y_test)

# test set으로 predict한 post.prob
logreg.predict_proba(x_test)

# predict confidence score
logreg.decision_function(x_test[:5])

logreg.predict(x_test[:5])
