==> Python/ML_basic/4.Logistic Regression/7.softmax_classifier_with_numpy.ipynb <==
import sklearn
import numpy as np

from sklearn.datasets import load_iris
datasets = load_iris()

x_data = datasets["data"]
x_data[:5]

y_data = datasets["target"]
y_data

y_data = y_data.reshape([-1,1])   # 2-dim array로 변환
y_data[:3]

from sklearn.preprocessing import OneHotEncoder
enc = OneHotEncoder()
enc.fit(y_data)  
y_data = enc.transform(y_data).toarray()
y_data[-1]

from sklearn.preprocessing import MinMaxScaler

min_max_scaler = MinMaxScaler()
x_data_minmax = min_max_scaler.fit_transform(x_data)
x_data_minmax[:3]

x_0 = np.ones(x_data_minmax.shape[0])
x_data_minmax = np.column_stack((x_0, x_data_minmax))   # np.column_stack(): np.hstack()과 같음

x_data_minmax[:3]

weights = np.random.uniform(size=(3,5))
weights

def softmax(z):
    e = np.exp(z)
    p = e / np.sum(np.exp(z), axis=1).reshape([-1,1])
    return p

z = x_data_minmax.dot(weights.T)
z.shape

softmax(z[:5])

np.sum(softmax(z[:5]), axis=1)   # softmax이기 때문에 각 행의 합은 1이다

def cross_entropy_function(y, x, weights):
    z = x_data_minmax.dot(weights.T)
    result = - np.sum(
                np.sum(
                    (y * np.log(softmax(z))), axis=1).reshape((-1,1))
                )
    return result 

cross_entropy_function(y_data,x_data_minmax,weights)

def minimize_grdient(y, x, initial_weights, iterations = 500000, alpha=0.001):
    cost_history= []
    theta_history = []
    m = y.shape[0]
    theta = np.copy(initial_weights)
    
    number_of_classes = theta.shape[0]
    number_of_weights = theta.shape[1]
    
    for _ in range(iterations):
        original_theta = np.copy(theta)
        for k in range(number_of_classes):        
            for j in range(number_of_weights):
                partial_x = x[:, j]
                partial_entropy = y - softmax(x.dot(original_theta.T))
                theta[k][j]  = original_theta[k][j] + (
                            alpha * partial_entropy[:,k].dot(partial_x.T) ) / 150
        if (_ % 10000) == 0:
            print(cross_entropy_function(y, x, theta) / 150)
            cost_history.append(cross_entropy_function(y, x, theta))
    return theta, cost_history

# weights = minimize_grdient(y_data, x_data_minmax,weights)
theta, cost_history = minimize_grdient(y_data, x_data_minmax, weights)

rand_index = np.random.randint(0, 150, 30)
rand_index

y_pred = np.argmax(softmax(x_data_minmax[rand_index].dot(theta.T)), axis=1) 
y_pred

y_true = np.argmax(y_data[rand_index], axis=1)
y_true

y_pred == y_true

# Accuracy
sum(y_pred == y_true) / len(rand_index)
