==> Python/ML_basic/4.Logistic Regression/9.classification_service.ipynb <==
import pandas as pd
import numpy as np

data_source = "http://www.ats.ucla.edu/stat/data/binary.csv"
df = pd.read_csv(data_source)
df[:5]

y_data = df["admit"].values.reshape(-1,1)
x_data = df.ix[:,1:].values

y_data[:5], x_data[:5]

from sklearn import preprocessing # Min-Max Standardzation

min_max_scaler = preprocessing.MinMaxScaler()
x_data = min_max_scaler.fit_transform(x_data)

x_data[:5]

from sklearn import linear_model, datasets

logreg = linear_model.LogisticRegression(fit_intercept=True)
logreg.fit(x_data, y_data.ravel())

sum(logreg.predict(x_data) == y_data.ravel()) / len(y_data.ravel())

theta = np.append(logreg.intercept_, logreg.coef_.ravel())
theta

theta_file_name = "theta_bin.npy"
np.save(theta_file_name, theta)

np.load(theta_file_name)

min_max = np.vstack( (df.ix[:,1:].values.min(axis=0), df.ix[:,1:].values.max(axis=0)) )
min_max

min_max_file_name = "min_max.npy"
np.save(min_max_file_name, min_max)

np.load(min_max_file_name)

np.load(theta_file_name)

import pickle

f = open('logicmodel.pkl', 'wb')
f.seek(0)

pickle.dump(logreg,f )
f.close()

fr = open('./logicmodel.pkl', 'rb')
logreg_pickle = pickle.load(fr)

sum(logreg.predict(x_data) == logreg_pickle.predict(x_data)) / len(y_data.ravel())
