==> Python/ML_basic/4.Logistic Regression/10.pipeline_example.ipynb <==
import pandas as pd
import numpy as np

columns = ["id", "diagnosis"] + [str("r"+str(i)) for i in range(30)]
df = pd.read_csv("https://archive.ics.uci.edu/ml/machine-learning-databases/breast-cancer-wisconsin/wdbc.data", names=columns)
X = df.loc[:,columns[2:]].values
Y = df.loc[:,columns[1]].values

from sklearn.model_selection import train_test_split
x_train, x_test, y_train, y_test = train_test_split(X, Y, test_size=0.2, random_state=22)

x_train.shape, y_train.shape, x_test.shape, y_test.shape

### Pipeline이란?
- 전처리 객체(Scaler)와 모형 객체(model)을 하나로 합쳐서 사용할 수 있게 해줌
- method
    - `fit` : `fit` 함수를 호출하면 scaler 객체에서는 `fit_transform`이 내부적으로 호출되고 model 객체에서는 `fit` 함수가 호출된다.
    - `predict`: `predict` 함수를 호출하면 scaler 객체에서는 `transform`이 내부적으로 호출되고 model 객체에서는 `predict` 함수가 호출된다 

from sklearn.preprocessing import StandardScaler
from sklearn.linear_model import LogisticRegression
from sklearn.pipeline import Pipeline

pipe_lr = Pipeline(steps=[('scl', StandardScaler()),
                          ('clf', LogisticRegression())])

pipe_lr.fit(X=x_train, y=y_train)
pipe_lr.score(x_test, y_test)
